refactor(wpcarro/emacs/pkgs): port to buildEmacsPackage
Almost all these expressions already depend on depot. This is most
notably the case for all expressions that are exported to
github.com/wpcarro, so the current implementation of the package builds
does not actually allow Nix builds independently from depot.

Currently, they somewhat clumsily replicate the API of buildEmacsPackage
so they can be used with depot. This unfortunately introduces a hard
dependency on the emacsPackages alias in Nixpkgs which has been removed
recently (see [also]). To deal with this, we may want to change the
internal API of buildEmacsPackage or implement a shim. This is much
easier to do when it only needs to happen in a single, central place.

While I was at it, I moved all test suites into installCheckPhase which
just requires -f package-activate-all instead of building a separate
version of the package and a emacsWithPackages derivation.

We can also change passthru.meta to meta without any problems.

[also]: https://github.com/NixOS/nixpkgs/pull/456065#issuecomment-3476250570

Change-Id: Ic91766ee8f88f1786679f8d0bedc4bf7456f299c
Reviewed-on: https://cl.tvl.fyi/c/depot/+/13814
Autosubmit: sterni <sternenseemann@systemli.org>
Tested-by: BuildkiteCI
Reviewed-by: wpcarro <wpcarro@gmail.com>
Reviewed-by: sterni <sternenseemann@systemli.org>